написан огромный интеграционный тест на API книг, новая ручка получения книги, фикс ручки PATCH, подробное документирование DBManager, моки книг и авторов, новые фикстуры на пинг редиса и Taskiq, создание файлика helpers.py, который в дальнейшем будет содержать полезные функции

+++ src/api/books.py
@@ -12,7 +12,7 @@ from src.exceptions.files import (
     WrongCoverResolutionHTTPException,
     WrongFileExpensionHTTPException,
 )
-from src.api.dependencies import S3Dep, DBDep, UserRoleDep, authorize_and_return_user_id
+from src.api.dependencies import S3Dep, DBDep, authorize_and_return_user_id
 from src.schemas.books import (
     BookAdd,
     BookAddWithAuthorsTagsGenres,
@@ -26,7 +26,6 @@ from src.tasks.taskiq_tasks import async_render, async_delete_book
 from src.schemas.books_authors import BookAuthorAdd
 from src.models.books import BooksTagsORM
 from src.validation.files import FileValidator
-from src.config import settings
 
 router = APIRouter(prefix="/author", tags=["Авторы и публикация книг"])
 
@@ -89,35 +88,40 @@ async def edit_bood_data(
     genres_ids_in_db = [genre.id for genre in genres]
     tags_titles_in_db = [tag.title_tag for tag in tags]
     # Вычисление нужных и НЕ нужных нам жанров
-    new_genres = data.genres
-    same_els_genres = set(genres_ids_in_db) & set(new_genres)
-    genres_to_add = set(new_genres) - same_els_genres
-    genres_to_delete = set(genres_ids_in_db) - same_els_genres
+    if data.genres is not None:
+        new_genres = data.genres 
+        same_els_genres = set(genres_ids_in_db) & set(new_genres)
+        genres_to_add = set(new_genres) - same_els_genres
+        genres_to_delete = set(genres_ids_in_db) - same_els_genres
     # Вычисление нужных и НЕ нужных нам тегов
-    new_tags = data.tags
-    same_els_tags = set(tags_titles_in_db) & set(new_tags)
-    tags_to_add = set(new_tags) - same_els_tags
-    tags_to_delete = set(tags_titles_in_db) - same_els_tags
+    if data.tags is not None:
+        new_tags = data.tags 
+        same_els_tags = set(tags_titles_in_db) & set(new_tags)
+        tags_to_add = set(new_tags) - same_els_tags
+        tags_to_delete = set(tags_titles_in_db) - same_els_tags
 
-    data_to_add_genres = [
-        GenresBooksAdd(genre_id=gen_book_id, book_id=book_id)
-        for gen_book_id in genres_to_add
-    ]
-    data_to_add_tags = [
-        TagAdd(title_tag=tag_book_title, book_id=book_id)
-        for tag_book_title in tags_to_add
-    ]
+    if data.genres is not None:
+        data_to_add_genres = [
+            GenresBooksAdd(genre_id=gen_book_id, book_id=book_id)
+            for gen_book_id in genres_to_add
+        ]
+    
+    if data.tags is not None:
+        data_to_add_tags = [
+            TagAdd(title_tag=tag_book_title, book_id=book_id)
+            for tag_book_title in tags_to_add
+        ]
 
     # --- Изменение жанров внутри базы данных ---
-    if genres_to_delete:
+    if data.genres is not None and genres_to_delete:
         await db.books_genres.delete_bulk_by_ids(genres_to_delete, book_id=book_id)
-    if genres_to_add:
+    if data.genres is not None and genres_to_add:
         await db.books_genres.add_bulk(data_to_add_genres)
-    if tags_to_delete:
+    if data.tags is not None and tags_to_delete:
         await db.tags.delete(
             BooksTagsORM.title_tag.in_(tags_to_delete), book_id=book_id
         )
-    if tags_to_add:
+    if data.tags is not None and tags_to_add:
         await db.tags.add_bulk(data_to_add_tags)
     if book_patch_data.model_dump(exclude_unset=True):
         await db.books.edit(data=book_patch_data, is_patch=True, book_id=book_id)
@@ -150,6 +154,16 @@ async def get_my_books(
     user_id: int = authorize_and_return_user_id(2),
 ):
     return await db.users.get_books_by_user(user_id=user_id)
+
+
+@router.get("/book")
+async def get_book_by_id(
+    book_id: int, db: DBDep
+) : 
+    try:
+        return await db.books.get_book_with_rels(book_id=book_id)
+    except BookNotFoundException as ex: 
+        raise BookNotFoundHTTPException
 
 
 # --- Обложки ---
